Fix the image loading error caused by path mismatch
This commit fixes the NullPointerException thrown when attempting to run the application outside IntelliJ (i.e. the CLI).
Though IJ required a '/src/' prefix, the compiled output also contained this prefix which did not correspond to the correct
path in the 'out/' directory as 'src/' is not included, only 'main/*'. It now looks for the resource path through the current
classes path, and navigates upward to the resources directory, in which it locates the given path to find the images.