Updates to client-server communication enabling successful comms.
This commit adds successful client-server communication of varying types. It will send the required information about the local Car object to the server, which will store it privately within that thread. To mock the remote client, the server will relay that same data back to the sending client when it requests the remote client Car information. This means the remote car will mock the actions of the singly-controlled local car on that client.

- Client and Server: Add new DTO for serialized communication of car properties to reduce request/response cycle complexity and count
- Client and Server: Migrate to use abstraction of Object streams to enable request/response of the multiple types required

- Client: Add preliminary "both players ready" request to server before loading main game screen (requires further implementation)
- Server: Lay groundwork for routing to other server thread to pass information to other client.

To Do:
- Add closure of Client network streams and notify server of closure, to relay to other client which must also handle and close.