Update to working communication of cars
This is a dirty, most likely non-thread safe solution which will be
refactored. Consider this the "make it work" stage of "make it work,
make it right, make it fast". There is a slight 1.5-2s delay on the Red
car only. Green is almost instant both locally and remotely, however,
Red suffers a delay for a currently unknown reason.