Resolve performance issues and inject sprites to track
Performance issues were caused by a number of factors. Firstly, the first client to connect to the server would bombard its message queue with objects. Thus, when the opposing client attempted to read them, it was having to go through a number of older objects first which appeared as latency. To resolve this, a restriction was placed upon the sending of local objects to only send if they have changed.

This introduced a further problem that upon sending there first and only update before being moved, the opposing client would retrieve it. At which point, the queue for that sprite would be emptied and the next request would result in a default object, and since the local sprite had not moved, it would not send another update. Queue starvation! Thus, both local sprites had to be moving in order to move remote sprites. To resolve this, a restriction was placed upon server-side queues to only perform a read and remove when more than 1 object exists in the queue, else only peek. Therefore, the objects are always up to date with their latest messages, and the queue is never fully empty thus leaving the corresponding client starved.